+++ python/triton/core.py
@@ -1,9 +1,246 @@
-from typing import Callable, TypeVar
+from __future__ import division, annotations
 
-from .base import *
+from typing import TypeVar, Callable
+
 from . import base
+from .base import tensor
 
 C = TypeVar("C", bound=Callable)
+
+
+@base.builtin
+def program_id(axis, _builder: base.ir.builder = None):
+    """
+    Returns the id of the current program instance along the given :code:`axis`.
+
+    :param axis: The axis of the 3D launch grid. Has to be either 0, 1 or 2.
+    :type axis: int
+    """
+    # if axis == -1:
+    #     pid0 = program_id(0, _builder)
+    #     pid1 = program_id(1, _builder)
+    #     pid2 = program_id(2, _builder)
+    #     npg0 = num_programs(0, _builder)
+    #     npg1 = num_programs(0, _builder)
+    #     return pid0 + pid1*npg0 + pid2*npg0*npg1
+    axis = base._constexpr_to_value(axis)
+    return base._program_id(axis, _builder)
+
+
+@base.builtin
+def num_programs(axis, _builder: base.ir.builder = None):
+    """
+    Returns the number of program instances launched along the given :code:`axis`.
+
+    :param axis: The axis of the 3D launch grid. Has to be either 0, 1 or 2.
+    :type axis: int
+    """
+    axis = base._constexpr_to_value(axis)
+    return base._num_programs(axis, _builder)
+
+
+@base.builtin
+def arange(start, end, _builder: base.ir.builder = None):
+    """
+    Returns contiguous values within the open interval [:code:`start`, :code:`end`).
+
+    :param start: Start of the interval. Must be a power of two.
+    :type start: int
+    :param stop: End of the interval. Must be a power of two >= start.
+    :type stop: int
+    """
+    start = base._constexpr_to_value(start)
+    end = base._constexpr_to_value(end)
+    return base._arange(start, end, _builder)
+
+
+@base.builtin
+def zeros(shape, dtype, _builder: base.ir.builder = None):
+    """
+    Returns a tensor filled with the scalar value 0 for the given :code:`shape` and :code:`dtype`.
+
+    :param shape: Shape of the new array, e.g., (8, 16) or (8, )
+    :type shape: tuple of ints
+    :param dtype: Data-type of the new array, e.g., :code:`float16`
+    :type dtype: DType
+    """
+    for i, d in enumerate(shape):
+        if not isinstance(d, base.constexpr):
+            raise TypeError(f"Shape element {i} must have type `constexpr`")
+        if not isinstance(d.value, int):
+            raise TypeError(
+                f"Shape element {i} must have type `constexpr[int]`, got `constexpr[{type(d.value)}]"
+            )
+    shape = [x.value for x in shape]
+    dtype = base._constexpr_to_value(dtype)
+    return base._zeros(shape, dtype, _builder)
+
+
+@base.builtin
+def dequantize(
+    input,
+    scale,
+    shift,
+    nbit,
+    dst_ty=base.float16,
+    _builder: base.ir.builder = None,
+):
+    """
+    Tries to dequantize the input to given dtype
+    """
+    nbit = base._constexpr_to_value(nbit)
+    return base._dequantize(input, scale, shift, nbit, dst_ty, _builder)
+
+
+@base.builtin
+def broadcast(input, other, _builder: base.ir.builder = None):
+    """
+    Tries to broadcast the two given blocks to a common compatible shape.
+
+    :param input: The first input tensor.
+    :type input: Block
+    :param other: The second input tensor.
+    :type other: Block
+    """
+    return base._broadcast_impl_value(input, other, _builder)
+
+
+@base.builtin
+def broadcast_to(input, shape, _builder: base.ir.builder = None):
+    """
+    Tries to broadcast the given tensor to a new :code:`shape`.
+
+    :param input: The input tensor.
+    :type input: Block
+    :param shape: The desired shape.
+    :type shape: Tuple[int]
+    """
+    return base._broadcast_impl_shape(input, shape, _builder)
+
+
+@base.builtin
+def cat(input, other, _builder: base.ir.builder = None):
+    """
+    Concatenate the given blocks
+
+    :param input: The first input tensor.
+    :type input:
+    :param other: The second input tensor.
+    :type other:
+    """
+    return base._cat(input, other, _builder)
+
+
+@base.builtin
+def reshape(input, shape, _builder: base.ir.builder = None):
+    """
+    Tries to reshape the given tensor to a new shape.
+
+    :param input: The input tensor.
+    :type input:
+    :param shape: The desired shape.
+    :type shape: Tuple[int]
+
+    """
+    shape = [x.value for x in shape]
+    return _reshape(input, shape, _builder)
+
+
+@base.builtin
+def dot(
+    input,
+    other,
+    trans_a=False,
+    trans_b=False,
+    allow_tf32=True,
+    _builder: base.ir.builder = None,
+):
+    """
+    Returns the matrix product of two blocks.
+
+    The two blocks must be two dimensionals and have compatible inner dimensions.
+
+    :param input: The first tensor to be multiplied.
+    :type input: 2D tensor of scalar-type in {:code:`float16`, :code:`bfloat16`, :code:`float32`}
+    :param other: The second tensor to be multiplied.
+    :type other: 2D tensor of scalar-type in {:code:`float16`, :code:`bfloat16`, :code:`float32`}
+    """
+    allow_tf32 = base._constexpr_to_value(allow_tf32)
+    return base._dot(input, other, trans_a, trans_b, allow_tf32, _builder)
+
+
+@base.builtin
+def load(
+    pointer,
+    mask=None,
+    other=None,
+    cache_modifier="",
+    eviction_policy="",
+    volatile=False,
+    _builder: base.ir.builder = None,
+) -> tensor:
+    """
+    Return a tensor of data whose values are, elementwise, loaded from memory at location defined by :code:`pointer`.
+
+    :param *:
+    :code:`mask` and :code:`other` are implicitly broadcast to :code:`pointer.shape`.
+
+    :code:`other` is implicitly typecast to :code:`pointer.dtype.element_ty`.
+
+    :param pointer: Pointers to the data to be loaded.
+    :type pointer: Block of dtype=triton.PointerDType
+    :param mask: if mask[idx] is false, do not load the data at address :code:`pointer[idx]`.
+    :type mask: Block of triton.int1, optional
+    :param other: if mask[idx] is false, return other[idx]
+    :type other: Block, optional
+    :param cache_modifier: changes cache option in nvidia ptx
+    'type cache_modifier: str, optional
+    """
+    # mask, other can be constexpr
+    if mask is not None:
+        mask = base._to_tensor(mask, _builder)
+    if other is not None:
+        other = base._to_tensor(other, _builder)
+    cache_modifier = base._constexpr_to_value(cache_modifier)
+    eviction_policy = base._constexpr_to_value(eviction_policy)
+    volatile = base._constexpr_to_value(volatile)
+    return base._load(
+        pointer,
+        mask,
+        other,
+        cache_modifier,
+        eviction_policy,
+        volatile,
+        _builder,
+    )
+
+
+@base.builtin
+def store(
+    pointer,
+    value,
+    mask=None,
+    eviction_policy="",
+    _builder: base.ir.builder = None,
+) -> tensor:
+    """
+    Stores :code:`value` tensor of elements in memory, element-wise, at the memory locations specified by :code:`pointer`.
+
+    :param *:
+    :code:`value` is implicitly broadcast to :code:`pointer.shape` and typecast to :code:`pointer.dtype.element_ty`.
+
+    :param pointer: The memory locations where the elements of :code:`value` are stored.
+    :type pointer: Block of dtype=triton.PointerDType
+    :param value: The tensor of elements to be stored.
+    :type value: Block
+    :param mask: If mask[idx] is false, do not store :code:`value[idx]` at :code:`pointer[idx]`.
+    :type mask: Block of triton.int1, optional
+    """
+    # value can be constexpr
+    value = base._to_tensor(value, _builder)
+    if mask is not None:
+        mask = base._to_tensor(mask, _builder)
+    return base._store(pointer, value, mask, eviction_policy, _builder)
 
 
 @base.builtin
@@ -275,10 +512,10 @@ def multiple_of(input, values, _builder: base.ir.builder = None):
     """
     Let the compiler knows that the values in :code:`input` are all multiples of :code:`value`.
     """
-    if isinstance(values, constexpr):
+    if isinstance(values, base.constexpr):
         values = [values]
     for i, d in enumerate(values):
-        if not isinstance(d, constexpr):
+        if not isinstance(d, base.constexpr):
             raise TypeError(f"values element {i} must have type `constexpr`")
         if not isinstance(d.value, int):
             raise TypeError(
@@ -293,10 +530,10 @@ def max_contiguous(input, values, _builder: base.ir.builder = None):
     """
     Let the compiler knows that the `value` first values in :code:`input` are contiguous.
     """
-    if isinstance(values, constexpr):
+    if isinstance(values, base.constexpr):
         values = [values]
     for i, d in enumerate(values):
-        if not isinstance(d, constexpr):
+        if not isinstance(d, base.constexpr):
             raise TypeError(f"values element {i} must have type `constexpr`")
         if not isinstance(d.value, int):
             raise TypeError(
@@ -311,12 +548,12 @@ def max_contiguous(input, values, _builder: base.ir.builder = None):
 # -----------------------
 
 
-@jit
+@base.jit
 def abs(x):
     return where(x >= 0, x, -x)
 
 
-@jit
+@base.jit
 def cdiv(x, div):
     """
     Computes the ceiling division of :code:`x` by :code:`div`
@@ -329,7 +566,7 @@ def cdiv(x, div):
     return (x + div - 1) // div
 
 
-@jit
+@base.jit
 def minimum(x, y):
     """
     Computes the element-wise minimum of :code:`x` and :code:`y`.
@@ -342,7 +579,7 @@ def minimum(x, y):
     return where(x < y, x, y)
 
 
-@jit
+@base.jit
 def maximum(x, y):
     """
     Computes the element-wise maximum of :code:`x` and :code:`y`.
@@ -355,22 +592,22 @@ def maximum(x, y):
     return where(x > y, x, y)
 
 
-@jit
+@base.jit
 @_add_math_1arg_docstr("sigmoid")
 def sigmoid(x):
     return 1 / (1 + exp(-x))
 
 
-@jit
+@base.jit
 @_add_math_1arg_docstr("softmax")
-def softmax(x, ieee_rounding: constexpr = constexpr(False)):
+def softmax(x, ieee_rounding: base.constexpr = base.constexpr(False)):
     z = x - max(x, 0)
     num = exp(z)
     den = sum(num, 0)
     return fdiv(num, den, ieee_rounding)
 
 
-@jit
+@base.jit
 def ravel(x):
     """
     Returns a contiguous flattened view of :code:`x`
@@ -381,7 +618,7 @@ def ravel(x):
     return reshape(x, [x.numel])
 
 
-@jit
+@base.jit
 def swizzle2d(i, j, size_i, size_j, size_g):
     """
     transformes indices of a row-major size_i*size_j matrix into those
@@ -414,6 +651,6 @@ def swizzle2d(i, j, size_i, size_j, size_g):
     return new_i, new_j
 
 
-@jit
+@base.jit
 def zeros_like(input):
     return zeros(input.shape, input.dtype)
