+++ python/triton/ops/cross_entropy.py
@@ -1,6 +1,8 @@
 import torch
 
 import triton
+import triton.core
+import triton.tuning
 
 tl = triton
 
@@ -23,49 +25,49 @@ def num_warps(N):
     return 16
 
 
-@triton.heuristics({'num_warps': lambda nargs: num_warps(nargs['N'])})
-@triton.heuristics({'BLOCK': lambda nargs: next_power_of_2(nargs['N'])})
+@triton.tuning.heuristics({'num_warps': lambda nargs: num_warps(nargs['N'])})
+@triton.tuning.heuristics({'BLOCK': lambda nargs: next_power_of_2(nargs['N'])})
 @triton.jit
 def _forward(LOGITS, PROBS, IDX, LOSS, N, BLOCK: tl.constexpr):
-    row = tl.program_id(0)
-    cols = tl.arange(0, BLOCK)
-    idx = tl.load(IDX + row)
+    row = triton.core.program_id(0)
+    cols = triton.core.arange(0, BLOCK)
+    idx = triton.core.load(IDX + row)
     # pointers to logit and probs
     LOGITS = LOGITS + row * N + cols
     WRIT_PROBS = PROBS + row * N + cols
     READ_PROBS = PROBS + row * N + idx
     # write-back negative log-probs
-    logits = tl.load(LOGITS, mask=cols < N, other=-float('inf'))
+    logits = triton.core.load(LOGITS, mask=cols < N, other=-float('inf'))
     logits = logits.to(tl.float32)
     logits = logits - tl.max(logits, 0)
     probs = tl.log(tl.sum(tl.exp(logits), 0)) - logits
-    tl.store(WRIT_PROBS, value=probs, mask=cols < N)
+    triton.core.store(WRIT_PROBS, value=probs, mask=cols < N)
     # There is a bug in the compiler, which fails to insert a barrier here.
     # We add it explicitly for now. Will be fixed soon.
     tl.debug_barrier()
     # write-back loss
-    probs = tl.load(READ_PROBS)
-    tl.store(LOSS + row, value=probs)
+    probs = triton.core.load(READ_PROBS)
+    triton.core.store(LOSS + row, value=probs)
 
 
-@triton.heuristics({'num_warps': lambda nargs: num_warps(nargs['N'])})
-@triton.heuristics({'BLOCK': lambda nargs: next_power_of_2(nargs['N'])})
+@triton.tuning.heuristics({'num_warps': lambda nargs: num_warps(nargs['N'])})
+@triton.tuning.heuristics({'BLOCK': lambda nargs: next_power_of_2(nargs['N'])})
 @triton.jit
 def _backward(PROBS, IDX, DPROBS, N, BLOCK: tl.constexpr):
-    row = tl.program_id(0)
-    cols = tl.arange(0, BLOCK)
-    idx = tl.load(IDX + row)
+    row = triton.core.program_id(0)
+    cols = triton.core.arange(0, BLOCK)
+    idx = triton.core.load(IDX + row)
     # pointers to probs
     PROBS = PROBS + row * N + cols
     # We know d(-log(p[i])/dlogit[k] = -id_mat[i,k] + p[k]
     # and we have -log(p[k]) stored in PROBS, so this is easy
-    probs = -tl.load(PROBS, mask=cols < N, other=float('inf'))
+    probs = -triton.core.load(PROBS, mask=cols < N, other=float('inf'))
     probs = tl.exp(probs.to(tl.float32))
     delta = cols == idx
     # write result in-place in PROBS
-    dout = tl.load(DPROBS + row)
+    dout = triton.core.load(DPROBS + row)
     din = (probs - delta) * dout
-    tl.store(PROBS, value=din.to(PROBS.dtype.element_ty), mask=cols < N)
+    triton.core.store(PROBS, value=din.to(PROBS.dtype.element_ty), mask=cols < N)
 
 
 class _cross_entropy(torch.autograd.Function):
